+++ lib/src/config/config.dart
@@ -21,12 +21,12 @@ class XDefaultThemeConfig extends XThemeConfig {
   List<XTheme> get themes => [
         XTheme(
             id: 'light',
-            nameBuilder: () => 'Light',
+            name: 'Light',
             theme: ThemeData.light(),
             colorScheme: LightColorScheme()),
         XTheme(
             id: 'dark',
-            nameBuilder: () => 'Dark',
+            name: 'Dark',
             theme: ThemeData.dark(),
             colorScheme: DarkColorScheme()),
       ];