+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.5
+- [Breaking change] : Refactor `XThemeConfig` [nameBuilder] function to String [name]
+- Added [updateById]: Updates theme using it's id.
+- Added [forceUpdateTheme] variable to change theme functions which force the app to rebuild all widget after changing the theme. Useful when depending on colorScheme in your widgets.
+
+
 ## 0.0.4
 - Add support for Dart 3.0.0 and Flutter 3.10
 - Add new `XColorScheme` which is a new way to configure custom colors for each themes.